migrate prisma to main repo

+++ src/server/endpoints/savedDecks/savedDecks.ts
@@ -1,6 +1,8 @@
 import { Request, Response, Express } from 'express';
 
 import { Prisma, PrismaClient, SavedDeck } from '@prisma/client';
+import { checkJwt } from '../../authz/checkJWT';
+import { getUserFromJWT } from '../../authz/getUserFromJWT';
 import { EmptyObj, ErrorMessage, SuccessMessage } from '@/types';
 
 export const initializeSavedDeckEndpoints = (
@@ -9,6 +11,7 @@ export const initializeSavedDeckEndpoints = (
 ) => {
     server.get(
         '/api/saved_decks/:username',
+        checkJwt,
         async (
             req: Request<
                 { username: string },
@@ -22,6 +25,18 @@ export const initializeSavedDeckEndpoints = (
             if (!username)
                 return res.status(400).send({ message: 'Need a username' });
 
+            const user = await getUserFromJWT(req.auth?.token);
+
+            if (!user) {
+                return res.status(400).send({ message: 'Need a username' });
+            }
+
+            if (user.username !== username) {
+                return res.status(401).send({
+                    message: `Hey! You can not see another user's decks!`,
+                });
+            }
+
             const savedDecks = await prisma.savedDeck.findMany({
                 where: { user: { username } },
             });
@@ -31,13 +46,14 @@ export const initializeSavedDeckEndpoints = (
 
     server.post(
         '/api/saved_decks',
+        checkJwt,
         async (
             req: Request<
                 EmptyObj,
                 SavedDeck | ErrorMessage,
                 {
                     deckName: string;
-                    skeleton: Prisma.JsonArray;
+                    skeleton: Prisma.JsonObject;
                     username: string;
                 }
             >,
@@ -49,10 +65,26 @@ export const initializeSavedDeckEndpoints = (
                 return res.status(400).send({ message: 'Need a username' });
             if (!deckName)
                 return res.status(400).send({ message: 'Need a deck name' });
-            if (!skeleton || !Array.isArray(skeleton))
+            if (
+                !skeleton ||
+                !Array.isArray(skeleton.mainBoard) ||
+                !Array.isArray(skeleton.sideBoard)
+            )
                 return res
                     .status(400)
                     .send({ message: 'Need a deck skeleton in JSON form' });
+
+            const userFromAuth0 = await getUserFromJWT(req.auth?.token);
+
+            if (!userFromAuth0) {
+                return res.status(400).send({ message: 'Need a username' });
+            }
+
+            if (userFromAuth0.username !== username) {
+                return res.status(401).send({
+                    message: `Hey! You can not create another user's decks!`,
+                });
+            }
 
             const user = await prisma.user.findFirst({ where: { username } });
             if (!user)
@@ -71,8 +103,76 @@ export const initializeSavedDeckEndpoints = (
         }
     );
 
+    server.patch(
+        '/api/saved_decks',
+        checkJwt,
+        async (
+            req: Request<
+                EmptyObj,
+                SavedDeck | ErrorMessage,
+                { deckId: string; skeleton: Prisma.JsonObject }
+            >,
+            res: Response<SavedDeck | ErrorMessage>
+        ): Promise<Response<SavedDeck | ErrorMessage>> => {
+            const { deckId, skeleton } = req.body;
+
+            if (!deckId)
+                return res.status(400).send({ message: 'Need a deck id' });
+            if (
+                !skeleton ||
+                !Array.isArray(skeleton.mainBoard) ||
+                !Array.isArray(skeleton.sideBoard)
+            )
+                return res
+                    .status(400)
+                    .send({ message: 'Need a deck skeleton in JSON form' });
+
+            const userFromAuth0 = await getUserFromJWT(req.auth?.token);
+
+            if (!userFromAuth0) {
+                return res.status(400).send({ message: 'Need a username' });
+            }
+
+            const user = await prisma.user.findFirst({
+                where: { username: userFromAuth0.username },
+            });
+            if (!user)
+                return res
+                    .status(400)
+                    .send({ message: 'No matching user found' });
+
+            const deck = await prisma.savedDeck.findFirst({
+                where: {
+                    id: deckId,
+                },
+            });
+
+            if (!deck) {
+                return res
+                    .status(400)
+                    .send({ message: 'No matching deck found' });
+            }
+            if (deck.userUid !== user.uid) {
+                return res
+                    .status(401)
+                    .send({ message: "Cannot update another user's deck!" });
+            }
+
+            const newDeck = await prisma.savedDeck.update({
+                where: {
+                    id: deckId,
+                },
+                data: {
+                    skeleton,
+                },
+            });
+            return res.send(newDeck);
+        }
+    );
+
     server.delete(
         '/api/saved_decks',
+        checkJwt,
         async (
             req: Request<
                 EmptyObj,
@@ -87,8 +187,19 @@ export const initializeSavedDeckEndpoints = (
                 return res.status(400).send({ message: 'Need a username' });
             if (!deckId)
                 return res.status(400).send({ message: 'Need a deck id' });
-            // TODO: add unit tests
-            // TODO: add auth0 layer
+
+            const userFromAuth0 = await getUserFromJWT(req.auth?.token);
+
+            if (!userFromAuth0) {
+                return res.status(400).send({ message: 'Need a username' });
+            }
+
+            if (userFromAuth0.username !== username) {
+                return res.status(401).send({
+                    message: `Hey! You can not delete another user's decks!`,
+                });
+            }
+
             try {
                 await prisma.savedDeck.deleteMany({
                     where: {
@@ -121,8 +232,19 @@ export const initializeSavedDeckEndpoints = (
 
             if (!username)
                 return res.status(400).send({ message: 'Need a username' });
-            // TODO: add unit tests
-            // TODO: add auth0 layer
+
+            const userFromAuth0 = await getUserFromJWT(req.auth?.token);
+
+            if (!userFromAuth0) {
+                return res.status(400).send({ message: 'Need a username' });
+            }
+
+            if (userFromAuth0.username !== username) {
+                return res.status(401).send({
+                    message: `Hey! You can not delete another user's decks!`,
+                });
+            }
+
             try {
                 await prisma.savedDeck.deleteMany({
                     where: {
