+++ src/server/endpoints/users/users.ts
@@ -1,7 +1,22 @@
 import { Request, Response, Express } from 'express';
 
 import { Prisma, PrismaClient, User } from '@prisma/client';
+import { checkJwt } from '../../authz/checkJWT';
+import { getUserFromJWT } from '../../authz';
+import { LEVELS } from '../../consts/xpSystem';
 import { EmptyObj, ErrorMessage, SuccessMessage } from '@/types';
+
+interface NewUserRequestBodyParams {
+    uid: string;
+    username: string;
+}
+
+interface ChooseAvatarBodyParams {
+    avatarUrl: string;
+}
+
+const DEFAULT_AVATAR =
+    'https://monksandmages.com/images/units/alert-feline.webp';
 
 export const initializeUserEndpoints = (
     server: Express,
@@ -9,6 +24,7 @@ export const initializeUserEndpoints = (
 ) => {
     server.get(
         '/api/users',
+        checkJwt,
         async (
             _: Request<EmptyObj, User[], EmptyObj>,
             res: Response<User[]>
@@ -18,10 +34,91 @@ export const initializeUserEndpoints = (
         }
     );
 
-    interface NewUserRequestBodyParams {
-        uid: string;
-        username: string;
-    }
+    server.get(
+        '/api/users/self',
+        checkJwt,
+        async (
+            req: Request<EmptyObj, User, EmptyObj>,
+            res: Response<User | ErrorMessage>
+        ) => {
+            const user = await getUserFromJWT(req.auth?.token);
+
+            if (!user) {
+                return res.status(400).send({ message: 'Need a username' });
+            }
+
+            const foundUser = await prisma.user.findUnique({
+                where: {
+                    username: user.username,
+                },
+            });
+
+            if (!foundUser) {
+                return res.status(204).send({ message: 'No user found' });
+            }
+
+            return res.send(foundUser);
+        }
+    );
+
+    server.patch(
+        '/api/users/self/choose_avatar',
+        checkJwt,
+        async (
+            req: Request<EmptyObj, User, ChooseAvatarBodyParams>,
+            res: Response<User | ErrorMessage>
+        ) => {
+            const user = await getUserFromJWT(req.auth?.token);
+            const { avatarUrl } = req.body;
+            if (!avatarUrl) {
+                return res
+                    .status(400)
+                    .send({ message: 'Need a avatarUrl field' });
+            }
+
+            if (!user) {
+                return res.status(400).send({ message: 'Need a username' });
+            }
+
+            const foundUser = await prisma.user.findUnique({
+                where: {
+                    username: user.username,
+                },
+            });
+
+            if (!foundUser) {
+                return res.status(204).send({ message: 'No user found' });
+            }
+
+            if (avatarUrl !== DEFAULT_AVATAR) {
+                const matchingLevel = LEVELS.find(
+                    (level) => avatarUrl === level.image
+                );
+                if (!matchingLevel) {
+                    return res
+                        .status(204)
+                        .send({ message: 'avatarUrl not in available images' });
+                }
+                if (foundUser.exp < matchingLevel.xpRequired) {
+                    return res.status(204).send({
+                        message:
+                            'avatarUrl not eligible - not high enough of a level',
+                    });
+                }
+            }
+
+            await prisma.user.update({
+                where: {
+                    username: user.username,
+                },
+                data: {
+                    avatarUrl,
+                },
+            });
+
+            return res.send(foundUser);
+        }
+    );
 
     server.post(
         '/api/users/new_user',
@@ -38,8 +135,14 @@ export const initializeUserEndpoints = (
                 return res
                     .status(400)
                     .send({ message: 'Need both a username and a uid' });
-            // TODO: add unit tests
-            // TODO: add auth0 layer
+
+            const apiKey = req.header('x-api-key');
+            if (apiKey !== process.env.API_KEY) {
+                return res.status(401).send({
+                    message: 'Not authorized!  Missing the right API key',
+                });
+            }
+
             try {
                 const user = await prisma.user.create({
                     data: { username, uid },
@@ -51,7 +154,8 @@ export const initializeUserEndpoints = (
                     error.code === 'P2002'
                 ) {
                     return res.status(400).send({
-                        message: 'A user with this uid already exists',
+                        message:
+                            'A user with this uid already exists OR this username is already taken',
                     });
                 }
 
@@ -75,8 +179,14 @@ export const initializeUserEndpoints = (
             const { uid } = req.body;
 
             if (!uid) return res.status(400).send({ message: 'Need a uid' });
-            // TODO: add unit tests
-            // TODO: add auth0 layer
+
+            const apiKey = req.header('x-api-key');
+            if (apiKey !== process.env.API_KEY) {
+                return res.status(401).send({
+                    message: 'Not authorized!  Missing the right API key',
+                });
+            }
+
             try {
                 const deleteUser = prisma.user.delete({
                     where: {
